pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["hyun <hyunsokstar@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "~4.0"
django-environ = "^0.9.0"
djangorestframework = "^3.14.0"
pyjwt = "^2.6.0"
django-summernote = "^0.8.20.0"
django-cors-headers = "^3.13.0"
dj-database-url = "^1.2.0"
psycopg2-binary = "^2.9.5"
brotli = "^1.0.9"
django-brotli = "^0.2.1"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
